requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyYAML==5.4.1
requests==2.26.0
Shapely==1.7.1
urllib3==1.26.6
Werkzeug==2.0.1
